refactor: move error serialize for circular import
